PDJB-261: split step out into 2 parts
