+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/journeys/acceptOrRejectJointLandlordInvitation/steps/DeleteInvitationAndTokenStepConfig.kt
@@ -7,44 +7,17 @@ import uk.gov.communities.prsdb.webapp.journeys.Destination
 import uk.gov.communities.prsdb.webapp.journeys.JourneyStep
 import uk.gov.communities.prsdb.webapp.journeys.acceptOrRejectJointLandlordInvitation.AcceptOrRejectJointLandlordInvitationJourneyState
 import uk.gov.communities.prsdb.webapp.journeys.shared.Complete
-import uk.gov.communities.prsdb.webapp.journeys.shared.YesOrNo
-import uk.gov.communities.prsdb.webapp.models.viewModels.emailModels.JointLandlordInvitationRejectionEmail
-import uk.gov.communities.prsdb.webapp.services.AbsoluteUrlProvider
-import uk.gov.communities.prsdb.webapp.services.EmailNotificationService
 import uk.gov.communities.prsdb.webapp.services.JointLandlordInvitationService
 
 @JourneyFrameworkComponent
 class DeleteInvitationAndTokenStepConfig(
     private val invitationService: JointLandlordInvitationService,
     private val invitationRepository: JointLandlordInvitationRepository,
-    private val rejectionEmailSender: EmailNotificationService<JointLandlordInvitationRejectionEmail>,
-    private val absoluteUrlProvider: AbsoluteUrlProvider,
 ) : AbstractInternalStepConfig<Complete, AcceptOrRejectJointLandlordInvitationJourneyState>() {
     override fun mode(state: AcceptOrRejectJointLandlordInvitationJourneyState): Complete = Complete.COMPLETE
 
     override fun afterStepIsReached(state: AcceptOrRejectJointLandlordInvitationJourneyState) {
         val invitation = invitationService.getInvitationForJourney(state.journeyId)
-
-        if (state.acceptOrRejectStep.outcome == YesOrNo.NO) {
-            val propertyAddress = invitation.registeredOwnership.address.singleLineAddress
-            val propertyRecordUrl =
-                absoluteUrlProvider.buildPropertyDetailsUri(invitation.registeredOwnership.id).toString()
-
-            invitationService.addRejectedPropertyAddressToSession(propertyAddress)
-
-            invitation.registeredOwnership.landlords.forEach { landlord ->
-                rejectionEmailSender.sendEmail(
-                    landlord.email,
-                    JointLandlordInvitationRejectionEmail(
-                        recipientName = landlord.name,
-                        inviteeEmail = invitation.invitedEmail,
-                        propertyAddress = propertyAddress,
-                        propertyRecordUrl = propertyRecordUrl,
-                    ),
-                )
-            }
-        }
-
         invitationRepository.delete(invitation)
 
         val token = invitationService.getInvitationTokenForJourneyIdFromSession(state.journeyId)
