+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/services/JointLandlordInvitationService.kt
@@ -98,8 +98,7 @@ class JointLandlordInvitationService(
         invitingLandlord: Landlord,
     ): String {
         val invitation =
-            invitationRepository
-                .findById(invitationId)
+            invitationRepository.findById(invitationId)
                 .orElseThrow { ResponseStatusException(HttpStatus.NOT_FOUND, "Invitation not found") }
 
         if (invitation.registeredOwnership.id != propertyOwnership.id) {
